refactor(tools): implement strict cohesive class coordinator pattern per #2917
